Fix IE8 fontface, update usemin, add useminprepare workaround, update Readme
I updated several NPM packages to fix an issue with font revving that was preventing fonts showing in IE8. However this created a new issue with useminprepare and multiple files. The workaround is to create a new useminprepare.html which references all of the assets to be revved by usemin/useminprepare.